+++ app/src/main/java/net/gsantner/markor/util/DocumentIO.java
@@ -25,7 +25,6 @@ import net.gsantner.markor.R;
 import net.gsantner.markor.activity.MainActivity;
 import net.gsantner.markor.format.TextFormat;
 import net.gsantner.markor.format.markdown.MarkdownTextConverter;
-import net.gsantner.markor.format.zimwiki.ZimWikiHighlighterPattern;
 import net.gsantner.markor.model.Document;
 import net.gsantner.opoc.util.FileUtils;
 
@@ -36,8 +35,6 @@ import java.nio.charset.StandardCharsets;
 import java.security.SecureRandom;
 import java.util.Locale;
 import java.util.UUID;
-import java.util.regex.Matcher;
-import java.util.regex.Pattern;
 
 import other.de.stanetz.jpencconverter.JavaPasswordbasedCryption;
 import other.de.stanetz.jpencconverter.PasswordStore;
@@ -130,7 +127,7 @@ public class DocumentIO {
                 document.setFormat(TextFormat.FORMAT_KEYVALUE);
             } else if (TextFormat.CONVERTER_MARKDOWN.isFileOutOfThisFormat(fnlower)) {
                 document.setFormat(TextFormat.FORMAT_MARKDOWN);
-            } else if (isZimWikiFile(fnlower, document)) {
+            } else if (TextFormat.CONVERTER_ZIMWIKI.isFileOutOfThisFormat(document.getFile().getAbsolutePath())) {
                 document.setFormat(TextFormat.FORMAT_ZIMWIKI);
             } else {
                 document.setFormat(TextFormat.FORMAT_PLAIN);
@@ -149,16 +146,6 @@ public class DocumentIO {
             AppSettings.appendDebugLog("\n\n\n--------------\nLoaded document, filepattern " + document.getFile().getName().replaceAll(".*\\.", "-") + ", chars: " + c.length() + " bytes:" + c.getBytes().length + "(" + FileUtils.getReadableFileSize(c.getBytes().length, true) + "). Language >" + Locale.getDefault().toString() + "<, Language override >" + AppSettings.get().getLanguage() + "<");
         }
         return document;
-    }
-
-    private static boolean isZimWikiFile(String filename, Document document) {
-        return filename.endsWith(".txt") && containsZimWikiHeader(document);
-    }
-
-    private static boolean containsZimWikiHeader(Document document) {
-        Pattern headerPattern = ZimWikiHighlighterPattern.ZIMHEADER.pattern;
-        Matcher headerMatcher = headerPattern.matcher(document.getContent());
-        return headerMatcher.find();
     }
 
     public static synchronized boolean saveDocument(final Document document, final String text, final ShareUtil shareUtil, Context context) {
